Removed preparevm from the view, removed unnecesaries init  and created the mainScreenVM
