+++ PVPCPlanner/App/PVPCPlannerApp.swift
@@ -4,14 +4,14 @@ import SwiftUI
 struct PVPCPlannerApp: App {
     // register app delegate for Firebase setup
     @UIApplicationDelegateAdaptor(AppDelegate.self) var delegate
-
+    let mainScreenViewModel = MainScreenViewModel()
     var body: some Scene {
         WindowGroup {
-            MainScreen()
+            MainScreen(mainScreenViewModel: mainScreenViewModel)
         }
     }
 }
 
 #Preview {
-    MainScreen()
+    MainScreen(mainScreenViewModel: MainScreenViewModel())
 }
