+++ PVPCPlanner/View/Screen/MainScreen.swift
@@ -2,11 +2,11 @@ import SwiftUI
 
 struct MainScreen: View {
     let settingsViewModel = SettingsViewModel()
-    @State private var pricesViewModel: PricesVM?
+    let mainScreenViewModel: MainScreenViewModel
 
     var body: some View {
         TabView {
-            PricesView(vm: createPricesVM())
+            PricesView(vm: mainScreenViewModel.createPricesVM())
                 .tabItem {
                     Label("", systemImage: "eurosign.circle")
                 }
@@ -24,16 +24,6 @@ struct MainScreen: View {
     }
 }
 
-func createPricesVM() -> PricesVM {
-    let container = PVPCDatabaseContainer.shared.container
-    let dataSource = PVPCLocalDataSource(container: container)
-    return PricesVM(
-        getPricesUseCase: GetPricesUseCase(repository: NetworkRepository()),
-        addPVPCTOLocalDBUseCase: AddPVPCToLocaDBUseCase(dataSource: dataSource),
-        getPVPCByDayFromLocalDBUseCase: GetPVPCByDayFromLocalDBUseCase(dataSource: dataSource)
-    )
-}
-
 #Preview {
-    MainScreen()
+    MainScreen(mainScreenViewModel: MainScreenViewModel())
 }
