+++ PVPCPlanner/View/Screen/PricesView.swift
@@ -3,16 +3,11 @@ import SwiftUI
 
 struct PricesView: View {
     @AppStorage(AppStorageKeys.LOCATION.rawValue) var selectedLocation: Locations = .MainlandAndIslands
-
-    private var vm: PricesVM
-
-    init(vm: PricesVM) {
-        self.vm = vm
-    }
+    var vm: PricesVM
 
     var body: some View {
         VStack {
-            CustomLazyList(spacing: 30, listDirection: .vertical, backgroundColor: Color.white) {
+            CustomLazyList(spacing: 30, listDirection: .vertical, backgroundColor: Color.clear) {
                 ForEach(vm.prices, id: \.self) { price in
                     PricesCard(pvpcCardModel: price.toPVPCCardModel(location: selectedLocation))
                 }
