+++ PVPCPlanner/ViewModel/MainScreenViewModel.swift
@@ -1,0 +1,14 @@
+import Foundation
+
+@Observable
+class MainScreenViewModel {
+    func createPricesVM() -> PricesVM {
+        let container = PVPCDatabaseContainer.shared.container
+        let dataSource = PVPCLocalDataSource(container: container)
+        return PricesVM(
+            getPricesUseCase: GetPricesUseCase(repository: NetworkRepository()),
+            addPVPCTOLocalDBUseCase: AddPVPCToLocaDBUseCase(dataSource: dataSource),
+            getPVPCByDayFromLocalDBUseCase: GetPVPCByDayFromLocalDBUseCase(dataSource: dataSource)
+        )
+    }
+}
